implementata norma a tramite kernel e cambiato tipo matrici a float per avere atomic add

+++ CUDA/gramschmidt.h
@@ -34,5 +34,5 @@
 #  endif
 # endif /* !N */
 
-using DATA_TYPE = double;
+using DATA_TYPE = float;
